# Fix significant performance issue of fixed-step stochastic jump solver