This fixes a major performance issue with our fixed-step jump solvers. On CPU, the replacement of `jax.lax.cond` by `jax.lax.select` (see e.g. https://github.com/jax-ml/jax/issues/7934) leads to a x100 speedup for trajectories simulation of a qubit.

| Number of trajectories | Duration before the PR         | Duration after the PR | Speedup |
|---------------|---------------------|----------------|-|
| 1,000    |      12.7 s ± 1.36 s  | 119 ms ± 8.87 ms    | x106            |
| 10,000           | 158 s (1 run)      |       1.1 s ± 186 ms          |  x144 |
| 100,000           | ??     | 13.1 s ± 1.54 s per loop        | ??|

The new perfs on GPU are quite impressive:
| Number of trajectories | Duration after the PR (GPU) |
|---------------|---------------------|
| 1,000 | 27.5 ms ± 258 μs per loop |
| 10,000 | 50 ms ± 137 μs per loop |
| 100,000 | 185 ms ± 609 μs per loop |

<details>
  <summary>Benchmark code</summary>

```python
import dynamiqs as dq
import numpy as np
import jax

H = dq.zeros(2)
jump_ops = [dq.sigmam(), dq.sigmap()]
psi0 = dq.excited()
tsave = np.linspace(0, 2.0, 101)
key = jax.random.PRNGKey(42)
method = dq.method.EulerJump(dt=1e-3)
for ntrajs in [int(1e3), int(1e4)]:
    keys = jax.random.split(key, ntrajs)

    def foo():
        return dq.jssesolve(
            H, jump_ops, psi0, tsave, keys, method=method
        ).states.block_until_ready()

    %timeit -r1 -n1 -q foo()
    %timeit foo()
```

</details>

There is another speedup to implement for `jssesolve` vs `jsmesolve`, the same as #911. I can do it once this PR is merged.